remove unusef endpoitns
