+++ internal/handler/v1/router.go
@@ -8,17 +8,34 @@ import (
 )
 
 func RegisterRoutes(r *router.Router) {
-	AddCustomRoutes(r)
-	AddProxyRoutes(r)
-	AddPriceRoutes(r)
-	AddERC20Routes(r)
-	AddUtilsRoutes(r)
-	AddNetworkRoutes(r)
-	AddValidatorDirectoryRoutes(r)
-	AddAirtableRoutes(r)
-	AddTransactionRoutes(r)
-	AddBalancesRoutes(r)
-	AddStakingRoutes(r)
-	AddGovernanceRoutes(r)
-	AddValidatorsRoutes(r)
+	r.GET("/RemainingEpochs", RemainingEpochs)
+	r.GET("/VoteRecord/{chain}/{proposal_id}/{address}", VoteRecord)
+	r.GET("/Epochs/{chain}", Epochs)
+	r.GET("/BalanceByDenom/{chain}/{address}/{denom:*}", BalanceByDenom)
+	r.GET("/TxStatus/{chain}/{tx_hash}", TxStatus)
+	r.POST("/broadcast", Broadcast)
+	r.POST("/simulate", Simulate)
+	r.GET("/ERC20ModuleBalance", ERC20ModuleEmptyBalance)
+	r.GET("/ERC20ModuleBalance/{evmos_address}/{eth_address}", ERC20ModuleBalance)
+	r.GET("/isIBCExecuted/{tx_hash}/{chain}", isIBCExecuted)
+	r.GET("/NetworkConfig", NetworkConfig)
+	r.GET("/NetworkConfig/{name}", NetworkConfigByName)
+	r.GET("/Announcements", GetAnnouncements)
+	r.POST("/broadcastEip712", BroadcastMetamask)
+	r.POST("/ibcTransfer", IBCTransfer)
+	r.POST("/convertCoin", ConvertCoin)
+	r.POST("/convertERC20", ConvertERC20)
+	r.POST("/broadcastAmino", BroadcastAmino)
+	r.POST("/delegate", Delegate)
+	r.POST("/undelegate", Undelegate)
+	r.POST("/redelegate", Redelegate)
+	r.POST("/rewards", Rewards)
+	r.POST("/vote", Vote)
+	r.POST("/cancelUndelegation", CancelUndelegation)
+	r.GET("/BalanceByNetworkAndDenom/{chain}/{token}/{address}", BalanceByNetworkAndDenom)
+	r.GET("/EVMOSIBCBalance/{chain}/{address}", EVMOSIBCBalance)
+	r.GET("/totalStakedByAddress/{address}", TotalStakingByAddress)
+	r.GET("/stakingInfo/{address}", StakingInfo)
+	r.GET("/V1Proposals", V1GovernanceProposals)
+	r.GET("/AllValidators", AllValidators)
 }
